making disconnecting a service from xendo more reliable

+++ rest/middlewares/xendo.js
@@ -190,6 +190,11 @@ const xendoRemoveServiceFromUser = (req, res, next) => {
     xendoSwipesCredentials,
     xendoUserServiceId
   } = res.locals;
+
+  if (xendoUserServiceId === null) {
+    return next();
+  }
+
   const url = xendoConfig.userServiceEndpoint + xendoUserServiceId;
 
   request.delete(url, {
